+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0-SNAPSHOT - 08/19/16
+- Added method to create Relation
+- Added JUnit test for `createRelation`
+- Added `ParseRelation` object
 - Added Relation Query
 
 ## 0-SNAPSHOT - 08/18/16